Character Movement Template (Unfinished)
I'm too tired to keep goin tonight so ima pick up on the control of the character in the morning with a cup of coffee. Stuck on the global data of the player. It should come out to be accesible from anywhere but there's probably a stupid little logic error that I missed somewhere.