Walking Animations Working Smoothly
The walking animations are now smooth by not using the main loop iteration counter. It is unreliable in classes other than 'Board.java'. The next step is weapon idle animations and then the animation for swinging the weapon. when swinging the weapon there has to be a rectangle around it created to detect which enemies are hit in which direction.